+++ src/main/java/com/ghbt/ghbt_starbucks/api/user_has_starbucks_card/dto/RequestStarbucksCard.java
@@ -1,4 +1,4 @@
-package com.ghbt.ghbt_starbucks.api.user_has_starbuckscard.dto;
+package com.ghbt.ghbt_starbucks.api.user_has_starbucks_card.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
